Conflicts:
	src/test/Test.FAKECore/FSIHelperSpecs.cs

+++ src/test/Test.FAKECore/FSIHelperSpecs.cs
@@ -19,8 +19,10 @@ namespace Test.FAKECore
             var outStream = new StringWriter(sbOut);
             Console.SetOut(outStream);
             Tuple<bool, Microsoft.FSharp.Collections.FSharpList<ProcessHelper.ConsoleMessage>> result;
+            
             try
             {
+                
                 result = FSIHelper.executeBuildScriptWithArgsAndReturnMessages(scriptFilePath, new string[] { }, useCache, false);
             }
             finally
@@ -40,7 +42,10 @@ namespace Test.FAKECore
                 Console.WriteLine(x.Message);
             }
             var messages = result.Item2.Where(x => !x.IsError).Select(x => x.Message);
-            return sbOut.ToString().Replace("\r\n", "\n").Replace("\r", "\n");
+            return 
+                sbOut.ToString()
+                .Replace("Running Buildscript: " + scriptFilePath, "")
+                .Replace("\n", "").Replace("\r", "");
         }
 
         static string Run(string script, string arguments, bool useCache)
@@ -79,33 +84,41 @@ namespace Test.FAKECore
                     var scriptHash =
                             FSIHelper.getScriptHash(new Tuple<string, string>[] { sc(scriptFilePath, "printf \"foobar\"") });
 
-                    var cacheFilePath = "./.fake/" + scriptFileName + "_" + scriptHash + ".dll";
+                    var cacheFilePath = Path.Combine(".", ".fake", scriptFileName + "_" + scriptHash + ".dll");
 
                     File.Exists(cacheFilePath).ShouldEqual(false);
 
-                    RunExplicit(scriptFilePath, arguments, false).ShouldEqual("foobar");
+                    RunExplicit(scriptFilePath, arguments, false)
+                       .ShouldEqual("foobar");
+
                     File.Exists(cacheFilePath).ShouldEqual(false);
 
                     RunExplicit(scriptFilePath, arguments, true)
-                        .ShouldEqual(("Cache doesnt exist" + nl + "foobar" + nl + "Saved cache" + nl).Replace("\r\n", "\n").Replace("\r", "\n"));
+                        .ShouldEqual(
+                            ("Cache doesnt exist" + nl + "foobar" + nl + "Saved cache" + nl)
+                            .Replace("\n", "").Replace("\r", ""));
+
                     File.Exists(cacheFilePath).ShouldEqual(true);
 
                     RunExplicit(scriptFilePath, arguments, true)
-                        .ShouldEqual(("Using cache" + nl + "foobar").Replace("\r\n", "\n").Replace("\r", "\n"));
+                        .ShouldEqual(
+                            ("Using cache" + nl + "foobar")
+                            .Replace("\n", "").Replace("\r", ""));
 
                     File.WriteAllText(scriptFilePath, "printf \"foobarbaz\"");
 
                     var changedScriptHash = FSIHelper.getScriptHash(new Tuple<string, string>[] { sc(scriptFilePath, "printf \"foobarbaz\"") });
                     RunExplicit(scriptFilePath, arguments, true)
                         .ShouldEqual(
-                        ("Cache is invalid, recompiling" + nl + "foobarbaz" + nl + "Saved cache" + nl).Replace("\r\n", "\n").Replace("\r", "\n"));
+                            ("Cache is invalid, recompiling" + nl + "foobarbaz" + nl + "Saved cache" + nl)
+                            .Replace("\n", "").Replace("\r", ""));
 
                     File.Exists("./.fake/" + scriptFileName + "_" + changedScriptHash + ".dll").ShouldEqual(true);
-
                 }
                 finally
                 {
-                    if (File.Exists(scriptFilePath)) File.Delete(scriptFilePath);
+                    if (File.Exists(scriptFilePath))
+                        File.Delete(scriptFilePath);
                 }
             };
 
@@ -120,18 +133,19 @@ namespace Test.FAKECore
                         "printf \"main\"\n#load \"" +
                             loadedPath.ToString().Replace("\\", "/") + "\"";
                     var loadedScript = "printf \"loaded;\"";
-                    File.WriteAllText(mainPath, mainScript.Replace("\r\n", "\n").Replace("\r", "\n"));
-                    File.WriteAllText(loadedPath, loadedScript.Replace("\r\n", "\n").Replace("\r", "\n"));
+                    File.WriteAllText(mainPath, mainScript.Replace("\r\n", "\n").Replace("\r", "\n").Replace("\r", nl));
+                    File.WriteAllText(loadedPath, loadedScript.Replace("\r\n", "\n").Replace("\r", "\n").Replace("\r", nl));
 
                     RunExplicit(mainPath, "", false)
                         .ShouldEqual("loaded;main");
                 }
                 finally
                 {
-                    File.Delete(mainPath);
-                    File.Delete(loadedPath);
+                    if (File.Exists(mainPath))
+                        File.Delete(mainPath);
+                    if (File.Exists(loadedPath))
+                        File.Delete(loadedPath);
                 }
-
             };
 
         It should_change_hash_when_loaded_file_changes =
